requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
google-cloud-firestore==2.16.0
pydantic==1.10.14
python-dotenv==1.0.1
twilio==9.2.3
pillow==10.3.0
pytesseract==0.3.10
openai==1.3.5
langid==1.1.6
numpy==2.3.3